requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
cryptography==43.0.1
bcrypt==4.2.0
PyJWT==2.9.0
pydantic[email]==2.9.2
reportlab==4.2.4
pandas==2.2.3
stripe==10.12.0
google-auth==2.35.0
python-dotenv==1.0.1
